implement Exercise4
