+++ src/components/App/App.js
@@ -2,13 +2,14 @@ import React from 'react';
 
 import ToastPlayground from '../ToastPlayground';
 import Footer from '../Footer';
+import ToastProvider from '../ToastProvider';
 
 function App() {
   return (
-    <>
+    <ToastProvider>
       <ToastPlayground />
       <Footer />
-    </>
+    </ToastProvider>
   );
 }
