+++ src/components/Toast/Toast.js
@@ -1,26 +1,19 @@
 import React from 'react';
 import {
-  AlertOctagon,
-  AlertTriangle,
-  CheckCircle,
-  Info,
   X,
 } from 'react-feather';
 
 import VisuallyHidden from '../VisuallyHidden';
 
 import styles from './Toast.module.css';
+import { useToast } from './hooks';
 
-const ICONS_BY_VARIANT = {
-  notice: Info,
-  warning: AlertTriangle,
-  success: CheckCircle,
-  error: AlertOctagon,
-};
-
-function Toast({ message, variant, id, onClickCloseToast }) {
-  const Icon = ICONS_BY_VARIANT[variant] || ICONS_BY_VARIANT['notice'];
-  const iconStyle = styles[variant] || styles['notice']
+function Toast({ message, variant, id }) {
+  const {
+    Icon,
+    iconStyle,
+    onClickCloseToast
+  } = useToast({ variant });
   return (
     <div className={`${styles.toast} ${iconStyle}`}>
       <div className={styles.iconContainer}>
