+++ src/components/ToastShelf/ToastShelf.js
@@ -2,13 +2,17 @@ import React from 'react';
 
 import Toast from '../Toast';
 import styles from './ToastShelf.module.css';
+import { useToastShelf } from './hooks';
 
-function ToastShelf({ toastList, onClickCloseToast }) {
+function ToastShelf() {
+  const {
+    toastList,
+  } = useToastShelf();
   return (
     <ol className={styles.wrapper}>
       {toastList.map((toast) => (
         <li key={toast.id} className={styles.toastWrapper}>
-          <Toast variant={toast.variant} message={toast.message} id={toast.id} onClickCloseToast={onClickCloseToast} />
+          <Toast variant={toast.variant} message={toast.message} id={toast.id} />
         </li>
       ))}
     </ol>
