+++ src/components/ToastShelf/hooks.js
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { ToastContext } from '../ToastProvider';
+
+export function useToastShelf() {
+  const {
+    toastList,
+  } = useContext(ToastContext);
+
+  return {
+    toastList,
+  }
+}
